Improve no-object-declaration type check
