+++ test/rules/no_object_declaration/example.dart
@@ -1,0 +1,29 @@
+class Test {
+  Object data = 1;
+
+  Object get getter => 1;
+
+  Object doWork() {
+    return null;
+  }
+}
+
+class AnotherTest {
+  int data = 1;
+
+  int get getter => 1;
+
+  void doWork() {
+    return;
+  }
+
+  void doAnotherWork(Object param) {
+    return;
+  }
+}
+
+Object a = 1;
+
+Object function(Object param) {
+  return null;
+}
